Full API

Documentation for Beamlines.

Beamlines.AlignmentParams — Type

AlignmentParams

Describe the alignment of the element with respect to the nominal position. Rotations are applied in order: tilt, x_rot, y_rot.

Properties

x_offset: Offset along x-axis [m]

y_offset: Offset along y-axis [m]

z_offset: Offset along z-axis [m]

x_rot: Rotation around the x-axis [rad]

y_rot: Rotation around the y-axis [rad]

tilt: Rotation around the z-axis [rad]

Beamlines.ApertureParams — Type

ApertureParams

Describe a mechanical aperture.

Properties

x1_limit: Left x-axis aperture edge [m]

x2_limit: Right x-axis aperture edge [m]

y1_limit: Lower y-axis aperture edge [m]

y2_limit: Upper y-axis aperture edge [m]

aperture_shape: Shape of the aperture, see ApertureShape

aperture_at: Longitudinal aperture location(s), see ApertureAt

aperture_shifts_with_body: true if the aperture position moves with element alignment, false otherwise. Default is true

aperture_active: true if particles are collimated by aperture, false otherwise. Default is true

Beamlines.BMultipoleParams — Type

BMultipoleParams

Defines magnetic (B) multipoles for an element.

A given multipole can be set using any of the following:

KnX: Order X normal strength in [1/m^(X+1)] (normalized, nonintegrated)

KsX: Order X skew strength in [1/m^(X+1)] (normalized, nonintegrated)

BnX: Order X normal strength in [T/m^X] (unnormalized, nonintegrated)

BsX: Order X skew strength in [T/m^X] (unnormalized, nonintegrated)

KnXL: Order X normal strength in [1/m^X] (normalized, integrated)

KsXL: Order X skew strength in [1/m^X] (normalized, integrated)

BnXL: Order X normal strength in [T/m^(X-1)] (unnormalized, integrated)

BsXL: Order X skew strength in [T/m^(X-1)] (unnormalized, integrated)

tiltX: Tilt to only order X multipole

The solenoidal multipole can be set using any of the above with X = sol and omitting the n/s, e.g. Ksol, Bsol, KsolL, BsolL.

Note

The last set for a given multipole defines if both the normal and skew strengths for a given multipole order are normalized and integrated. E.g.

ele = LineElement(Kn1=0.1, Ks1=0.2, L=0.4) # Order 1 is normalized, nonintegrated
ele.Kn1L = 0.3         # Order 1 independent variables are now normalized, integrated
ele.Ks1L == 0.2*0.4    # true
ele.Bn1 = 0.5          # Order 1 independent variables are now unnormalized, nonintegrated
ele.Bs1 == 0.2*0.5/0.3 # true, set so that (Bs1/Bn1)_new = (Ks1L/Kn1L)_old

Beamlines.Beamline — Type

Beamline

Structure containing a vector of LineElements in an ordered sequence, and optionally a Context struct containing conrol variables associated with the beamline for convenience. The reference species (specified as species_ref and reference energy (specified as one of E_ref, pc_ref, p_over_q_ref, dE_ref, dpc_ref, or dp_over_q_ref) is uniform over the entire beamline. The most-recently specified of these reference energy quantites will be stored as the independent variable, in the first LineElement of the Beamline.

Properties

line: A read-only array of LineElements in the beamline, in order

context: Context struct containing control variables associated with the beamline. If in a Branch, this is shared by the whole Branch (and Lattice, if any), and setting it sets it for all of them

branch: Branch that the beamline is placed in, if any

branch_index: Index of the beamline in the Branch, if in a Branch

Beamlines.BeamlineParams — Type

BeamlineParams

Defines information for LineElements that are in a Beamline.

Properties

beamline: Beamline that this LineElement is in

beamline_index: Index of the line array of the Beamline that this element is at

s: Longitudinal position from the start of the Beamline at the entrance of the element [m]

s_downstream: Longitudinal position from the start of the Beamline at the exit of the element [m]

line: A read-only array of LineElements in the beamline, in order

context: Context struct containing control variables associated with the beamline. If in a Branch, this is shared by the whole Branch (and Lattice, if any), and setting it sets it for all of them

branch: Branch that the beamline is placed in, if any

branch_index: Index of the beamline in the Branch, if in a Branch

Beamlines.BendParams — Type

BendParams

Defines a curvature in the reference coordinate system throught the element, as well as the edge angles of the element at the entrance and exit of the element. Does not specify any magnetic field.

Properties

g_ref: Coordinate system curvature [1 / m]

tilt_ref: Tilt angle applied before the curvature, and undone after [rad]

e1: Edge angle of the entrance of the element w.r.t. the entering coordinate system [rad]

e2: Edge angle of the exit of the element w.r.t. the exit coordinate system [rad]

edge1_int: Pole face field integral at the entrance [T * m]

edge2_int: Pole face field integral at the exit [T * m]

Beamlines.Branch — Type

Branch

Structure containing a vector of Beamlines, where currently each follows in-order, one after the other.

Properties

name: Name of the Branch. Default is "" if not in a lattice and if in a lattice, the default is "bN" where N is the index of the branch in lattice.branches

beamlines: Vector of the beamlines in the Branch

lattice: Lattice that the branch is placed in, if any

lattice_index: Index of the branch in the Lattice, if in a Lattice

context: Context shared by the Branch and all of its Beamlines, and by the whole Lattice if the Branch is in one. Setting the context of any of them sets it for all of them.

Beamlines.Context — Type

Context{T}

A "scope" of variables that can be used when evaluating DefExprs. Contexts serve as a single structure that stores all (control) variables associated with beamline parameters. Contexts essentially act as structures with arbitrary field names. The type parameter T can be specified as the variables' return type, or Union of return types, to improve performance.

If a "get" is made of a variable not included in a given Context, then the "get" will fall-back to the global stack of contexts GLOBAL_CONTEXTS. The first instance from the top of the GLOBAL_CONTEXTS stack of the variable will be used as the variable value. This makes it easy to "push" and "pop" parameter settings one may be testing in an interactive environment.

Examples

julia> c1 = Context(a = 1, b = 2, c = 3);

julia> push!(GLOBAL_CONTEXTS, c1);

julia> c2 = Context(a = 4);

julia> d = DefExpr(c -> c.a + c.b);

julia> d() # Uses c1.a and c1.b from the GLOBAL_CONTEXTS
3

julia> d(c2) # Uses c2.a, but falls-back to c1.b from GLOBAL_CONTEXTS
6

julia> push!(GLOBAL_CONTEXTS, c2);

julia> d() # Now uses c2.a and c1.b both from GLOBAL_CONTEXTS
6

julia> pop!(GLOBAL_CONTEXTS);

julia> d() 
3

Beamlines also store a context, which is passed to all DefExprs when getting parameters from LineElements that are in a beamline:

julia> c3 = Context(Kn1=0.36);

julia> qf = Quadrupole(Kn1=DefExpr(c -> c.Kn1), L=0.5);

julia> bl = Beamline([qf], context=c3);

julia> bl[qf][1].Kn1
0.36

Notice that trying to access the value of Kn1 directly from qf does not work since c3 is not stored where it can be found:

julia> qf.Kn1
ERROR: Variable Kn1 is not defined in the local Context nor in GLOBAL_CONTEXTS

julia> push!(GLOBAL_CONTEXTS, c3)    # Make c3 available globally

julia> qf.Kn1                        # Now this works
0.36

Beamlines.Controller — Type

Controller

An eagerly-evaluated controller of LineElement properties or other Controller properties. Similar to classical Bmad's "Overlay" or "Group".

Also see set!.

Tip

In SciBmad, it is generally preferred to use the lazily-evaluated DefExpr for deferred expressions.

Examples

qf = Quadrupole(Kn1=0.36, L=0.5)
qd = Quadrupole(Kn1=-0.36, L=0.5)

c1 = Controller(
  (qf, :Kn1) => (ele; x) ->  x,
  (qd, :Kn1) => (ele; x) -> -x;
  vars = (; x = 0.0,)
)

# Now we can vary both simultaneously:
c1.x = 60.
qf.Kn1 == -qd.Kn1 == 60 # true

# Controllers also include the element itself. This can 
# be useful if the current elements' values should be 
# used in the function:
c2 = Controller(
  (qf, :Kn1) => (ele; dKn1) ->  ele.Kn1 + dKn1,
  (qd, :Kn1) => (ele; dKn1) ->  ele.Kn1 - dKn1;
  vars = (; dKn1 = 0.0,)
) 

c2.dKn1 = 20
qf.Kn1 == -qd.Kn1 == 80 # true

# Controllers can also be used to control other controllers:
c3 = Controller(
  (c1, :x) => (ctrl; dx) -> ctrl.x + dx;
  vars = (; dx = 0.0,)
)

Beamlines.DefExpr — Type

DefExpr{T}

A lazily-evaluated deferred expression returning type T. Deferred expressions are lambda functions that "close" over a variable in the current scope. They can be used to specify inter-dependent parameters in an accelerator and guarantee that no parameter ever becomes "stale".

Also see Context.

Examples

julia> a = 0.36;

julia> qf = Quadrupole(Kn1L=DefExpr(()->a)); # captures variable a

julia> qf.Kn1L
0.36

julia> a = 0.7
0.7

julia> qf.Kn1L
0.7

Slightly better performance may be achieved by explicitly typing the captured variable:

julia> b::Float64 = 0.2;

julia> qf = Quadrupole(Kn1L=DefExpr(()->b)); # captures variable b with known type

julia> qf.Kn1L
0.2

julia> b = 0.4
0.4

julia> qf.Kn1L
0.4

Deferred expressions can be treated and operated with as regular numbers, even outside the context of Beamlines, and can be evaluated by calling it like a function with no arguments (with ()):

julia> c = 64;

julia> cd = DefExpr(()->sin(c));

julia> c = pi;

julia> cd = DefExpr(()->sin(c));

julia> cd()
0.0

julia> c = pi/2;

julia> cd()
1.0

julia> dd = cd + 5;

julia> dd()
6.0

An optional Context argument can be provided:

julia> d = DefExpr(c -> c.a + c.b);

julia> c1 = Context(a = 1, b = 2);

julia> d(c1)
3

Beamlines.EMFieldParams — Type

EMFieldParams

Defines the electromagnetic field of an element as a function em_field(x, y, s, t, p=nothing) where p is a tuple of parameters that may be used inside the function.

Properties

em_field: Function with arguments (x, y, s, t, p) the returns a tuple of the EM field values (Ex, Ey, Ez, Bx, By, Bz) at the given point.

em_field_params: Tuple of parameters passed to the em_field function. Default is nothing

em_field_normalized: If true, then EM field values returned are normalized by p_over_q_ref; default is false.

Beamlines.EMultipoleParams — Type

EMultipoleParams

Defines electric (E) multipoles for an element.

A given multipole can be set using any of the following:

EnX: Order X normal strength in [V/m^(X+1)] (nonintegrated)

EsX: Order X skew strength in [V/m^(X+1)] (nonintegrated)

EnXL: Order X normal strength in [V/m^X] (integrated)

EsXL: Order X skew strength in [V/m^X] (integrated)

etiltX: Tilt to only order X multipole

Note

The last set for a given multipole defines if both the normal and skew strengths for a given multipole order are integrated. E.g.

ele = LineElement(En1=0.1, Es1=0.2, L=0.4) # Order 1 is nonintegrated
ele.En1L = 0.3         # Order 1 independent variables are now integrated
ele.Es1L == 0.2*0.4    # true

Beamlines.FourPotentialParams — Type

FourPotentialParams

Defines the electromagnetic four-potential of an element as a function four_potential(x, y, s, t, p=nothing) where p is a tuple of parameters that may be used inside the function.

Properties:

four_potential: Function with arguments (x, y, s, t, p=nothing) that returns a two component

tuple (C, D) with both C and D being themselves tuples:
⠀ C = (ϕ, Ax, Ay, As)
⠀ D = (∂ϕ/∂x, ∂ϕ/∂y, ∂ϕ/∂s, ∂ϕ/∂t,
⠀ ∂Ax/∂x, ∂Ax/∂y, ∂Ax/∂s, ∂Ax/∂t,
⠀ ∂Ay/∂x, ∂Ay/∂y, ∂Ay/∂s, ∂Ay/∂t,
⠀ ∂As/∂x, ∂As/∂y, ∂As/∂s, ∂As/∂t)
- four_potential_params: Tuple of parameters passed to the four_potential function. Default is nothing

four_potential_normalized: If true, then potential/derivatives returned are normalized by p_over_q_ref; default is false.

Beamlines.InheritParams — Type

InheritParams

Stores a parent LineElement through which any parameter groups NOT present in the "child" element containing the InheritParams would inherit. E.g., if the child element has its own BeamlineParams, then any property from the BeamlineParams parameter group would be read/ write to the child's BeamlineParams.

Properties:

parent: Parent LineElement to inherit parameter groups from for both reading and writing.

Beamlines.InitialBeamlineParams — Type

InitialBeamlineParams

Defines the reference species and energy of the beamline. These parameters may be "set" in the first element of the beamline or at the Beamline level, and retrieved at any element in the beamline. If the reference species or energy is not set, then it will inherit that quantity from an upstream beamline if in a Branch.

Properties

species_ref: Reference species of the beamline

E_ref: Total reference energy [eV]

pc_ref: Reference momentum [eV/c]

p_over_q_ref: A signed reference magnetic rigidity [T * m]

dE_ref: Change in total reference energy w.r.t. the directly-upstream beamline [eV]

dpc_ref: Change in reference momentum w.r.t. the directly-upstream beamline [eV/c]

dp_over_q_ref: Change in signed reference magnetic rigidty w.r.t. the directly-upstream beamline [T * m]

Beamlines.Lattice — Type

Lattice

Structure containing a vector of Branches.

Properties

name: Name of the Lattice. Defaults to blank "".

branches: Vector of the branches in the Lattice. These are copies of the Branches the Lattice was constructed with, which are left unmodified

context: Context shared by the Lattice, all of its Branches, and all of their Beamlines. Setting the context of any of them sets it for all of them.

Beamlines.Lattice — Method

Lattice(branches::Vector{Branch}; name = "", context = Context())

Constructs a Lattice given the vector of branches branches. The Lattice holds copies (see copy(::Branch)) of the Branches, whose LineElements are children of those of the corresponding Branch in branches. The Branches in branches are not modified, so a Branch may appear more than once and may also be used in other Lattices. Branches without a name are named "b<i>", where <i> is the index of the branch. The contexts of the Branches and context are merged into a single Context shared by the Lattice, all of its Branches, and all of their Beamlines. Variables in context take precedence over those in the Branches.

Example

bl1 = Beamline([Marker(E_ref=10e9, species_ref=Species("electron")), Drift(L=1)])
bl2 = Beamline([Drift(L=2)])

lattice = Lattice([Branch([bl1]), Branch([bl2])])

Lattice(beamlines::Vector{Beamline}; name = "", context = Context())

Constructs a Lattice containing a single Branch made up of the vector of Beamlines beamlines. As with Branch(beamlines), the Branch holds copies of the Beamlines and beamlines is not modified.

Example

bl1 = Beamline([Marker(E_ref=10e9, species_ref=Species("electron")), Drift(L=1)])
bl2 = Beamline([Drift(L=2)])

lattice = Lattice([bl1, bl2]) # Equivalent to Lattice([Branch([bl1, bl2])])

Beamlines.LineElement — Type

LineElement

The basic building block which makes up a Beamline. May be something physical like a quadrupole magnet, or something non-physical like a point in the accelerator you want to mark. A LineElement may define a region in space distinguished by the presence of (possibly time-varying) electromagnetic fields, materials, apertures and other possible properties.

LineElement properties are split into "parameter groups" for convenient organization:

  AlignmentParams         ApertureParams               BMultipoleParams
   x_offset                x1_limit                     KnX
   y_offset                x2_limit                     KsX
   z_offset                y1_limit                     BnX
   x_rot                   y2_limit                     BsX
   y_rot                   aperture_shape               KnXL
   tilt                    aperture_at                  KsXL
                           aperture_shifts_with_body    BnXL
                           aperture_active              BsXL
                                                        tiltX

  BeamlineParams          BendParams                   EMFieldParams
   beamline                g_ref                        em_field
   beamline_index          tilt_ref                     em_field_params
   s                       e1                           em_field_normalized
   s_downstream            e2
   line                    edge1_int
   context                 edge2_int
   branch
   branch_index

  EMultipoleParams        FourPotentialParams          InitialBeamlineParams
   EnX                     four_potential               species_ref
   EsX                     four_potential_params        E_ref
   EnXL                    four_potential_normalized    pc_ref
   EsXL                                                 p_over_q_ref
   etiltX                                               dE_ref
                                                        dpc_ref
                                                        dp_over_q_ref

  MapParams               MetaParams                   PatchParams
   transport_map           alias                        dt
   transport_map_params    label                        dx
                           description                  dy
                                                        dz
                                                        dx_rot
                                                        dy_rot
                                                        dz_rot

  RFParams                UniversalParams
   rf_frequency            kind
   harmon                  name
   phi0                    L
   zero_phase              tracking_method
   traveling_wave
   is_crabcavity

To set a property, use the natural syntax:

ele = LineElement()
ele.L = 1   # Length 
ele.Kn1 = 2 # Normal quadrupole strength
ele.rf_frequency = 1e6 

For detailed descriptions of properties in a given parameter group, see the documentation for that parameter group.

Beamlines.MapParams — Type

MapParams

Defines any arbitrary function to transport particle coordinates' and optionally spin quaternions. This could be a matrix, a neural network, a phase trombone, etc.

The input for the transport_map must be (v, q, p=nothing), where v is a tuple of the input phase space coordinates (x, px, y, py, z, pz) and q is nothing is there is no spin tracking, else q is a tuple of the input quaternion (q0, q1, q2, q3). p may be a tuple of parameters that can be used inside of the transport_map. This might be neural network weights, for example, or perhaps phase evolutions for a phase trombone.

For GPU compatibility, it is necessary that your function transport_map is GPU compatible: there should be no type instabilities. You should also try to make your transport_map be branchless, so both the CPU and GPU can vectorize it.

transport_map_params should contain the tuple that is passed into p at the time of tracking.

Properties:

transport_map: Arbitrary function that transports particles' coordinates v, and optionally spin quaternions q. Must have the signature <your map name>(v, q, p=nothing) and returns the output as a tuple (v, q)

transport_map_params: Parameters of the transport map that will be inputted as the p variable in transport_map

Example

using StaticArrays
random_matrix = @SMatrix rand(6,6) # use StaticArrays for GPU compatibility

function matrix(v, q::Nothing, p) # `q::Nothing` -> no spin tracking
  # Vectorizable matrix multiplication by `p`:
  vx  = sum(p[1,:] .* v)
  vpx = sum(p[2,:] .* v)
  vy  = sum(p[3,:] .* v)
  vpy = sum(p[4,:] .* v)
  vz  = sum(p[5,:] .* v)
  vpz = sum(p[6,:] .* v)
  return ((vx, vpx, vy, vpy, vz, vpz), q)
end

m = LineElement(transport_map=matrix, transport_map_params=random_matrix)

Beamlines.MetaParams — Type

MetaParams

Defines extra String properties that may be useful for pattern matching or storing extra information about a given element.

Properties

alias: Alternate name for the element as a string

label: A label string

description: A descriptive string

Beamlines.PatchParams — Type

PatchParams

Defines properties for patches, which change the reference coordinate system. Rotations are applied in order: dz_rot, dx_rot, dy_rot.

Properties

dt: Reference time shift [s]

dx: New coordinate system offset in the x-direction [m]

dy: New coordinate system offset in the y-direction [m]

dz: New coordinate system offset in the z-direction [m]

dx_rot: Rotation of coordinate system around the initial x-axis [rad]

dy_rot: Rotation of coordinate system around the initial y-axis [rad]

dz_rot: Rotation of coordinate system around the initial z-axis [rad]

Beamlines.RFParams — Type

RFParams

Defines parameters generally associated with radiofrequency (RF) cavities. The frequency of the oscillating electromagnetic field may be optionally specified using the property rf_frequency or harmon; whichever of these is set last will be the independent variable.

Properties

rf_frequency: Frequency of the oscillating electromagnetic field [Hz]

harmon: Harmonic number of the oscillating electromagnetic field w.r.t. the length of the entire containing Beamline [1]

phi0: Phase offset of the oscillating field w.r.t. zero_phase [rad]

zero_phase: A PhaseRef that specifies what phi0 should mean, see PhaseRef

traveling_wave: true if traveling wave, false if standing wave

is_crabcavity: true if this is a crab cavity, false otherwise

Beamlines.SciBmadStandard — Type

SciBmadStandard

Default tracking method that uses exact transport maps when solvable, else uses the symplectic integrator Yoshida(order=4, num_steps=1) which chooses an appropriate split for each element.

Properties

radiation_damping_on: true if the deterministic effect of synchrotron radiation is included, false otherwise. Defaults to false

radiation_fluctuations_on: true if the stochastic radiation kicks are included, false otherwise. Defaults to false

ibs_damping_on: true if the deterministic effect of intrabeam scattering (IBS) is included, false otherwise. Defaults to false.

ibs_fluctuations_on: true if the stochastic kicks of intrabeam scattering (IBS) is included, false otherwise. Defaults to false.

Beamlines.UniversalParams — Type

UniversalParams

Describes the kind, name, length, and tracking method for a LineElement.

Properties

kind: String specifing the "kind", of an element, e.g. "Quadrupole"

name: The name of an element as a string

L: Length of the element [m]

tracking_method: Tracking method for the element, defaults to SciBmadStandard()

Beamlines._Branch — Type

Internal: mutable struct _Branch{T<:_AbstractBeamline} <: _AbstractBranch

_Branch exists to break a mutual type recursion since Beamline has a reference to a Branch, and Branch needs a vector of Beamlines. Julia < v1.14 has no forward declarations, so to get around this, _Branch is used.

Why not just use Branch{T} and skip defining _Branch? This could be done but in this case Branch would become a UnionAll rather than a concrete type leading to type instability.

Base.copy — Method

Base.copy(bl::Beamline)

Copy of bl that is not in any Branch and has a copy of the Context of bl. Each LineElement of the copy is a new element that inherits (via InheritParams) from the corresponding element of bl and has its own BeamlineParams. Parameters are therefore shared with bl through inheritance, while the position of each element (s, beamline, etc.) is that in the copy. bl is not modified.

Base.copy — Method

Base.copy(branch::Branch)

Copy of branch that is not in any Lattice. The Beamlines are copied as with copy(::Beamline), so the LineElements of the copy are children of those in branch. The Context is copied.

Base.empty! — Method

empty!(::Beamline)

Removes BeamlineParams from all elements in the Beamline and empties the array of LineElements.

WARNING: this is irreversible.

Base.merge — Method

merge(c::Context, cs::Context...)

Construct a new Context containing the variables of all of the given Contexts. If a variable is defined in more than one of the Contexts, the value from the last one is used. The type parameter of the returned Context is the promote_type of the type parameters of the inputs, and each value is converted to it using coerce. For example, merging a Context{Float32} with a Context{Float64} gives a Context{Float64}. Only the variables stored in the given Contexts are included; variables from GLOBAL_CONTEXTS are not.

Examples

julia> c1 = Context(a = 1, b = 2);

julia> c2 = Context(c = 3, d = 4);

julia> c3 = merge(c1, c2);

julia> c3.a + c3.d
5

julia> c4 = merge(Context{Int}(a = 1), Context{Float64}(b = 2.0));

julia> typeof(c4)
Context{Float64}

julia> c4.a
1.0

Beamlines.BPM — Method

BPM(; kwargs...) = LineElement(; kind="BPM", kwargs...)

See the documentation for LineElement

Beamlines.CrabCavity — Method

CrabCavity(; kwargs...) = LineElement(; kind="CrabCavity", is_crabcavity = true, kwargs...)

See the documentation for LineElement

Beamlines.Drift — Method

Drift(; kwargs...) = LineElement(; kind="Drift", kwargs...)

See the documentation for LineElement

Beamlines.HKicker — Method

HKicker(; kwargs...) = LineElement(; kind="HKicker", kwargs...)

See the documentation for LineElement

Beamlines.Kicker — Method

Kicker(; kwargs...) = LineElement(; kind="Kicker", kwargs...)

See the documentation for LineElement

Beamlines.Marker — Method

Marker(; kwargs...) = LineElement(; kind="Marker", kwargs...)

See the documentation for LineElement

Beamlines.Multipole — Method

Multipole(; kwargs...) = LineElement(; kind="Multipole", kwargs...)

See the documentation for LineElement

Beamlines.Octupole — Method

Octupole(; kwargs...) = LineElement(; kind="Octupole", kwargs...)

See the documentation for LineElement

Beamlines.Patch — Method

Patch(; kwargs...) = LineElement(; kind="Patch", kwargs...)

See the documentation for LineElement

Beamlines.Quadrupole — Method

Quadrupole(; kwargs...) = LineElement(; kind="Quadrupole", kwargs...)

See the documentation for LineElement

Beamlines.RFCavity — Method

RFCavity(; kwargs...) = LineElement(; kind="RFCavity", kwargs...)

See the documentation for LineElement

Beamlines.SBend — Method

SBend(; kwargs...) = LineElement(; kind="SBend", kwargs...)

See the documentation for LineElement

Beamlines.Sextupole — Method

Sextupole(; kwargs...) = LineElement(; kind="Sextupole", kwargs...)

See the documentation for LineElement

Beamlines.Solenoid — Method

Solenoid(; kwargs...) = LineElement(; kind="Solenoid", kwargs...)

See the documentation for LineElement

Beamlines.VKicker — Method

VKicker(; kwargs...) = LineElement(; kind="VKicker", kwargs...)

See the documentation for LineElement

Beamlines._context — Method

_context(bl::Beamline)

Return the Context of bl: the Context stored in the Lattice or Branch that bl is in, if any, else the Context stored in bl itself. See _context(::Branch).

Beamlines._context — Method

_context(branch::Branch)

Return the Context of branch: the Context stored in the Lattice that branch is in, if any, else the Context stored in branch itself.

The Context of a Lattice, or of a Branch not in a Lattice, is stored only at that highest level. The Branches and Beamlines below it store NULL_CONTEXT, which is set when they are put in the Branch or Lattice. Setting the context property at any level sets the field of the highest level only.

Beamlines.elements — Method

elements(eles_dict)

For more Pythonic-workflows, elements receives a dictionary with key as the element name and value as the element, and sets the name attribute for each element equal to its key.

Julia users should generally use the @elements macro instead.

Beamlines.findchildren — Method

findchildren(ele::LineElement, bl::Beamline)

Finds all LineElements in the Beamline bl that inherit from ele, either directly (ele is the parent) or through a chain of parents. For example, the elements of a Beamline in a Branch are children of the elements of the Beamline used to create the Branch, which are in turn children of the elements used to create that Beamline.

Beamlines.scalarize! — Method

scalarize!(bl::Beamline)

Modifies the Beamline and its LineElements so all parameters are regular number types. This may be needed after optimizing the element's parameters using e.g. ForwardDiff, ReverseDiff, or GTPSA, which will set the parameter equal to a special number type that propagates the gradients.

Beamlines.scalarize! — Method

scalarize!(branch::Branch)

Modifies all Beamlines and their LineElements in the Branch so all parameters are regular number types. This may be needed after optimizing the element's parameters using e.g. ForwardDiff, ReverseDiff, or GTPSA, which will set the parameter equal to a special number type that propagates the gradients.

Beamlines.scalarize! — Method

scalarize!(ele::LineElement)

Modifies the LineElement so all element-level parameters are regular number types. This may be needed after optimizing the element's parameters using e.g. ForwardDiff, ReverseDiff, or GTPSA, which will set the parameter equal to a special number type that propagates the gradients.

Beamlines.set! — Method

set!(controller)

Resets all controlled LineElements and Controllers in controller to the most recently set values of that controller.

Beamlines.writebl — Method

writebl is internal, still being developed, and may change at any time. General users should NOT use it yet.

Beamlines.@elements — Macro

@elements ...;
@elements begin
  ...
end

Can be applied before the definition of any LineElement(s) in order to make set the element name equal to the variable symbol automatically.

Examples

julia> @elements qf = Quadrupole();

julia> qf.name
"qf"

julia> @elements begin
         my_drift = Drift(L=0.5);
         my_sextupole = Sextupole(Kn2L=3.2);
       end;

julia> my_drift.name
"my_drift"

julia> my_sextupole.name
"my_sextupole"

Beamlines.ApertureAt — Module

@enumx ApertureAt::UInt8 Entrance Exit BothEnds

Defines at which ends of the element the aperture should be applied

Beamlines.ApertureShape — Module

@enumx ApertureShape::UInt8 Elliptical Rectangular

Defines the shape of the aperture

Beamlines.PhaseRef — Module

@enumx PhaseRef::UInt8 Accelerating BelowTransition AboveTransition

Sets what zero phi0 RF phase means

Accelerating: Zero phase is the maximum accelerating phase.

BelowTransition: Zero phase is at the stable zero crossing for particles below transition.

AboveTransition: Zero phase is at the stable zero crossing for particles above transition.

Beamlines.RateMeaning — Module

@enumx RateMeaning::Int8 RFFrequency=false Harmon=true Indeterminate=-1

Specifies which independent variable is stored in the RFParams.